Tests implementation backup

+++ TicketSystemAPI/Functions/Tickets/Command/DeleteTicketCommandHandler.cs
@@ -1,23 +1,28 @@
 ﻿using MediatR;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.EntityFrameworkCore;
+using Microsoft.AspNetCore.Authorization;
 using System.Linq;
-using System.Net.Sockets;
 using System.Threading;
 using System.Threading.Tasks;
-using TicketSystemAPI;
+using TicketSystemAPI.Authorization;
 using TicketSystemAPI.Data;
 using TicketSystemAPI.Exceptions;
+using TicketSystemAPI.Services;
 
 namespace TaskSystemAPI.Functions.Tickets.Command
 {
     public class DeleteTicketCommandHandler : IRequestHandler<DeleteTicketCommand>
     {
         private readonly DataContext _context;
+        private readonly IUserContextService _userContextService;
+        private readonly IAuthorizationService _authorizationService;
 
-        public DeleteTicketCommandHandler(DataContext context)
+        public DeleteTicketCommandHandler(DataContext context, 
+                                            IUserContextService userContextService,
+                                            IAuthorizationService authorizationService)
         {
             _context = context;
+            _userContextService = userContextService;
+            _authorizationService = authorizationService;
         }
 
         public async Task<Unit> Handle(DeleteTicketCommand request, CancellationToken cancellationToken)
@@ -29,8 +34,13 @@ namespace TaskSystemAPI.Functions.Tickets.Command
             }
             try
             {
-                _context.Tickets.Remove(entity);
-                await _context.SaveChangesAsync(cancellationToken);
+                var authorizationResult = _authorizationService.AuthorizeAsync(_userContextService.User, entity, new ResourceOperationRequirement(ResourceOperation.Delete)).Result;
+                if (authorizationResult.Succeeded) // ToDo
+                {
+                    _context.Tickets.Remove(entity);
+                    await _context.SaveChangesAsync(cancellationToken);
+                }
+                throw new BadRequestException("Ticket was not removed.");
             }
             catch
             {
